Add type system and scalar user-defined function framework (#2)
This PR adds the framework required for Scalar UDFs (and implements
three of them to make sure it's going to work: `ST_(Geog)Point()`,
`ST_Geom/GeogFromText()`, and `ST_AsText()`). The functions I did
implement I tried to do in such a way that they don't cut corners on
documentation or testing (so that hopefully they are somewhat complete
examples to draw from when implementing new functions).

The basic idea is to ensure that the basic `invoke_batch()`
implementation can remain pretty much unchanged between the
proof-of-concept functions and here. The `SeondaPhysicalType` enum
manages wrapping, unwrapping, and type equality and currently just
supports WKB encoding. The `SimpleSedonaUdf` implements the wrapping and
unwrapping of arguments and includes the ability to register multiple
"kernels" based on a matching function (our version of a Signature that
incorporates some flexibility). I'm envisioning that FFI will be
implemented at the "kernel" level but it might be that it all folds into
the UDF eventually. When datafusion has support for user-defined types,
this part can all disappear anyway because the `Signature` will take
care of the matching and casting.

It will be a bit of an adventure to add other encodings here (maybe
Kristin's aligned Sedona serialization, maybe PostGIS gserialized, maybe
GeoArrow, maybe something Geography-specific), but Rust's `geo-traits`
are awesome and I think can form the basis for functions implemented in
Rust. Even if the current WKB implementation of the Geometry trait isn't
particularly fast (pretty sure it heap allocates unavoidably), we can
make it better or use the `geoarrow` implementation of the Geometry
trait or use FFI and fork off to something else once we have benchmarks
to check these things.

The tests here are pretty verbose and I'm hoping we can add some helpers
to improve on them. I do think we need these tests to be there, though,
because they will be the basis for the other test helpers and it's a bit
circular to test them with themselves.