feat(sedona-cli): Add basic CLI for testing (#6)
This is not intended to be a fully featured CLI, just something to take
care of the basic details of setting up filesystem configuration.

The CLI can be started with `cargo run`:

```
Sedona CLI v0.0.1
> SELECT ST_Point(0, 1) as geom;
+------------------------------------------------------------+
| geom                                                       |
+------------------------------------------------------------+
| {geoarrow.wkb: 01010000000000000000000000000000000000f03f} |
+------------------------------------------------------------+
1 row(s) fetched. 
Elapsed 0.002 seconds.
```

The current implementation is a straight vendor of
https://github.com/apache/datafusion/ 's `datafusion-cli`, which
unfortunately is not very pluggable.

There is also https://github.com/datafusion-contrib/datafusion-dft ,
which is probably closer to what we'd want if we ever want this for
non-developers.

(superceeds #4)