feat(rust/geoarrow-testing): Update geoarrow-data and add helper scripts (#28)
This PR updates geoarrow-data to the brand-new 0.2.0 release, and adds a
script to download the asset files. I also consolidated the helper
function to access the files in tests (that we can use in benchmarks,
too). In tests and forthcoming benchmarks you can now write:

```rust
test_geoparquet("natural-earth", "countries")
```

...and this should give you a reasonable error if you forgot to clone
with submodules or run the asset downloading script to make the bigger
files available.