feat(geoparquet) Add with_predicate and with_schema_adapter_factory (#31)
This PR is to add the following two implementations to the
GeoParquetFileSource struct:
- with_predicate: apply a predicate to the inner parquet file source if
supported
- with_schema_adapter_factory: apply a schema adapter factory to the
inner parquet file source if supported

Both of them are needed to support using the GeoParquet loader in Glue
when creating the DataSourceExec plan. See this PR:
https://github.com/wherobots/sedona-glue/pull/12

This PR does not implement applying/pushing geospatial filters down to
the reader, so when it is supported, the filters should also be applied
to geospatial columns as needed. Same for the schema adapter logic as
well, as it currently just pass to the inner parquet file source.