feat(rust/sedona-geometry): Add sedona-geometry crate with BoundingBox (#36)
This PR implements the incredibly boring math associated with combining
bounding boxes with full support for wraparound. This is needed for the
predicate pushdown into Parquet and didn't really exist anywhere else
(particularly with wraparound logic). Strictly speaking we could have
errored whenever we saw wraparound (it's technically supported in
GeoParquet, but probably rare), but we've punted on that a lot of times
and at least now there's a comprehensive set of tests for this.