fix(rust/sedona-geoparquet): Fix edge serialization when present (#37)
This fixes the output when edges were present in the GeoParquet
metadata. We need to go through and properly support non-lon/lat CRSes
(as well as go through and test this metadata code, since there were no
tests when I copied it from geoarrow-rs). It needs heavy modification
for filter pushdown and I'd prefer to defer tests until that point.