refactor(rust/sedona-geo,rust/sedona-functions): Move ST_Area() to stub function (#40)
Because we need to implement st_area in at least two crates (one for
geometry, one for geography), we need this as a stub in
sedona-functions. This PR moves the definition of `st_area_udf()` to
`sedona-functions`.

In general, we will define the actual implementations of functions in
crates specific to the dependency that implements them. For example, we
implement `st_intersects` for GEOMETRY using the Rust `geo` crate, so we
have a sedona-geo crate that provides all the implementations that
depend on `geo`. The implementation for GEOGRAPHY will depend on C++
`s2-geography`, and we'll assemble everything together in `sedona` (or
in the future, perhaps at runtime).

This also fixes a number of clippy warnings that appeared after the rust
release today. I think there's a better way to consolidate them but
`[workspace.lints.clippy]` didn't work out of the box and I'd prefer to
figure that out later if we need more linter configuration.