refactor(rust/sedona-expr): Remove unused arguments from SedonaScalarKernel trait (#41)
This PR removes the `return_types` and `num_rows` arguments from the
`SedonaScalarKernel` trait. As Feng pointed out, these aren't used
anymore and the `GenericExecutor` gives us a way to calculate the number
of output rows given arguments that might be scalars or arrays.

The implications of this for kernels is that we can no longer implement
zero-argument kernels using this framework. I think that's
ok...zero-argument kernels are pretty much just random generators, and
we don't need any framework to do this (a regular DataFusion `ScalarUDF`
would suffice).

This also means that if we have all Scalar inputs, we have to return all
scalar outputs. I think this is OK, too...it's good practice and I can't
think of a function that needs to take all scalar inputs and produce an
array output. If we do run across this constraint, we can add an
advanced version that passes along more from the scalar udf.