feat(rust/sedona-functions): Add ST_Length stub function (#43)
This PR is a template for adding a stub function, which is a function
definition in sedona-functions that exists so we can use multiple crates
to define implementations. Stub functions define the name, input, and
output types and documentation. The steps are:

- Pick a function. Lots of choices here:
https://docs.wherobots.com/latest/references/wherobotsdb/vector-data/Function/
- Find a similar function. I used `ST_Area()` as a template (i.e., I
copy/pasted the content of st_area.rs to start the process). This
function is similar because it has the same input and output types
(i.e., geometry or geography -> float64).
- Ensure the input and output types are correct (i.e., the `ArgMatcher`
is defined correctly
- Add your module to `lib.rs` (in this case `mod st_area;`)
- Ensure the documentation and example is up-to-date from the
docs.wherobots.com definition. In the future we'll make sure examples
run in CI, but that day is not this day.
- Add the function to the function set in `register.rs`.
- Ensure it builds and the "is it plugged in" test runs
- From a terminal, run `cargo run` and try your function in SQL. This
will give an error (because it's a stub function, not an
implementation), but you shouldn't get a `Error during planning: Invalid
function 'st_length'.`.
- Run `pre-commit run --all-files`. This will check for egregious typos
and Rust code formatting. You may have to set up a virtual environment
first and `pip install pre-commit` first.
- PR!

```
# cargo run
Sedona CLI v0.0.1
> SELECT ST_Length(ST_GeomFromWKT('LINESTRING (0 1, 2 3)'));
This feature is not implemented: Implementation for st_length([Wkb(Planar, None)]) was not registered
```