chore: Upgrade to DataFusion main (pre-48) (#51)
This PR updates the Arrow and DataFusion dependencies to a commit just
before the release of DataFusion 48 (which is the same commit currently
being used by Comet). DataFusion 48 should be released sometime in the
next two weeks. The main change is that predicate filters are supplied
to the data source in a new way, so to get our pruning we needed to
update some details about our wrapping of the Parquet implementation.

DataFusion 48 also includes the ability to propagate field metadata
which means we can remove the wrap/unwrap code. I think we should wait
until the final DataFusion 48 for that because there's at least one PR
that has a breaking change.