feat(python/sedona-rs): Basic DuckDB-ish Python interface (#45)
This PR adds the required infrastructure to the Python package to get us
to:

```python
import sedona_rs
con = sedona_rs.connect()

con.sql("SELECT 1 as one").to_arrow_table()
#> pyarrow.Table
#> one: int64 not null
#> ----
#> one: [[1]]
```

(Pandas output is also supported, using GeoPandas if there's any
geometry)

Basically, we need a `SessionContext` equivalent and a `DataFrame`. We
can make syntactic shortcuts to some of these things if needed (e.g.,
make a "default" context so that you can do `sedona_rs.sql()`), but
given that what we actually want is `sedona.sql()`, we might want to put
that in sedona instead.

I'd initially though that we should run everything through SQL (like
DuckDB), but this approach (more closely wrapping objects in
Rust/Python) is easier and we can figure out how to map some of these
concepts to SQL if we have to.