Fix the st_envelope aggr state signature to match Sedona UDAF (#59)
This PR is to fix the issue found in integration with the Glue, for
witch it failed due to the mismatch of the state variable. In Sedona
UDAF, we define the st_envelope_aggr as:
```
class ST_Envelope_Aggr
    extends Aggregator[Geometry, Geometry, Geometry]
    with TraitSTAggregateExec { ...
```
It dictate the state being a Geometry (binary) type and it will be used
in the CometExchange phase after the Partial aggregation phase and
before the Full aggregation phase.

As for currently in Comet, even if all phases (partial, exchange, full)
all happen in the native code, it still verifies the state in jni
driver. Therefore, we will need to match the state in aggregator with
the Sedona UDAF state variables. This change is needed to pass this
validation on the Comet side.

This also applies to other aggregate functions, intersection, union,
analyze, which the intermediate states would be complex objects.