[EWT-2755, EWT-2754] Spatial join using GEOS prepared geometry (#73)
This PR implements spatial join execution modes using GEOS prepared
geometry. The performance is terrible, so they are not enabled by
default. The problems were discussed in
[EWT-2755](https://www.notion.so/wherobots/In-memory-join-with-prepared-build-side-217ba32ec0278061969dde5b4f8824b6).
We'll address these problems in follow-up PRs.

This PR also contains some clean ups of existing code:

* Refactored index building into SpatialIndexBuilder.
* Keep geo Wkb objects with evaluated batches, so that it will be easier
for us to retrieve Wkb by index.
* Add sedona options extension, now we can configure spatial join
options using `SET sedona.* TO value` statement
* Fixed handling of set statements in SedonaContext

The effort for abstracting away different computational geometry
libraries is in another branch:
https://github.com/wherobots/sedona-db/tree/sjoin-multi-lib. This PR is
already large so I'd like to keep it as is. Support for multiple
geometry libraries will be in an upcoming PR.