feat(c/sedona-s2geography): Add sedona-s2geography crate (#70)
This PR adds a sedona-s2geography crate that builds and statically links
a copy of google/s2geometry and paleolimbot/s2geography. Linking system
OpenSSL and Abseil is a bit of a trick...this currently uses CMake to
locate them and passes a text file to `build.rs` that is parsed into the
values that Rust needs to link a target. This works on Mac and Linux
(provided CMake can find Abseil and OpenSSL via `find_package()`);
however, doesn't quite work on MSVC yet.

This seems to be cached effectively by cargo and hasn't added
significant build time for me but we can also use feature flags or
separate workspaces to ensure this doesn't affect the developer
experience of non-geography enthusiasts.

I'll follow up with separate PRs that actually implement features!