refactor(rust): Make aggregate functions slightly less verbose to test (#90)
This PR implements a new method `.aggregate_wkt()` to the
AggregateTester (that makes it less verbose to construct inputs) and a
new function `assert_scalar_equal_wkb_geometry()` that makes checking an
output geometry less verbose.

We don't have that many more aggregate functions to implement, but this
does shave 30% ish off of the aggregate function testing lines and does
make them a little less annoying to write.

I'll follow up with a `ScalarUdfTester` that will hopefully make things
nicer for scalar functions.