feat(rust/sedona-testing): Enable even more compact tests (#105)
This adds some `assert_xx()` functions to the ScalarUdfTester to make it
easier to scale our testing.

The failure mode could be improved by making these a macro but it also
might not matter. These are mostly useful for "is it plugged in" sort of
tests on functions whose actual behaviour is tested at a lower level
already (and whose behaviour at scale will be tested at a higher level).