Reimplement the Rust binding for PROJ and add CachingCrsEngine (#139)
This PR resolves the performance issue described in
https://www.notion.so/wherobots/Slow-ST_Transforms-with-Proj-250ba32ec0278092926de0374cbe6f67:

1. Reimplement the Rust proj binding based on proj-sys because the
original implementation is problematic: creating a dedicated proj context each
time we call proj functions is an anti-pattern, it will also call
`proj_cleanup` when some proj objects are still being used in
multi-threaded environment. We create a proj context for each thread
and reuse it instead. This also reuses the connection to the sqlite
database and make proj lookup faster.
2. Caches proj transforms. The cache lookup is still pretty time
consuming for now, but it is way faster than creating new proj
transforms.

https://github.com/wherobots/sedona-db/pull/134 is basically the same as
this PR, but contains changes made to st_transform.rs. It can be a
reference for integrating cache into ST_Transform in
https://github.com/wherobots/sedona-db/pull/127