chore: Add test geometry fixtures ported from the geo-test-fixtures crate in georust/geo (#193)
This is part of the forked dependency elimination plan: https://github.com/apache/sedona-db/pull/165.

We'll maintain our generic geo algorithms refactored from georust/geo, the test fixtures for checking the correctness of our refactored implementation is also needed. Unfortunately, geo-test-fixtures is not published to crates.io, so we have to copy it to our own projects to use them.

The WKT files were already committed to the apache/sedona-testing repository (see https://github.com/apache/sedona-testing/pull/9), which is a submodule of sedona-db. We add fixtures for loading the WKT files in the submodule to sedona-testing. These newly added fixtures will be used by future PRs.