feat: Add sedona-geo-traits-ext to sedona-db (#194)
This is part of the forked dependency elimination plan: https://github.com/apache/sedona-db/pull/165. This PR depends on https://github.com/apache/sedona-db/pull/193.

This PR moves geo-traits-ext from wherobots/geo to sedona-db and renamed it to sedona-geo-traits-ext. Currently, it is a standalone crate and can be compiled using `cd rust/sedona-geo-traits-ext && cargo build`. We'll update the Cargo.toml files in the final step to make it live.