chore: Make KnnComponents field optional in SpatialIndex (#437)
`KnnComponents` is only used when running KNN join (spatial predicate is `SpatialPredicate::KNearestNeighbors(...)`), and it reserves lots of memory for cached geometry objects. Currently the KnnComponents is always initialized even when running range join or distance join. This patch changes `knn_components` to be optional and only present when running KNN join.